chore: add roundtrip test
