+++ core/Cargo.toml
@@ -16,3 +16,4 @@ spargebra = { git = "https://github.com/jeswr/oxigraph.git", rev = "b32b1f3", fe
 sophia_jsonld = "0.9.0"
 serde_json = "1.0.139"
 thiserror = "2.0.11"
+serde_test = "1.0.177"
